# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIPD API Documentation",
        "description": "Dokumentasi API untuk aplikasi SIPD",
        "contact": {
            "email": "admin@sipd_new.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://sipd_new.test/api/",
            "description": "Local Server"
        }
    ],
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "operationId": "ed42f392010c301892e640067141676e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Response Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "properties": {
                                                "status": {
                                                    "type": "integer",
                                                    "example": 201
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "register success"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Joko Widodo"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "jokowi@gmail.com"
                                                },
                                                "nip_nik": {
                                                    "type": "string",
                                                    "example": "199001012020121004"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "081234567895"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "operationId": "8cb7742ae98d0990acb2907c4e00bf24",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Response Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "properties": {
                                                "status": {
                                                    "type": "integer",
                                                    "example": 200
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "login success"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "data": {
                                            "properties": {
                                                "permissions": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "create-request",
                                                        "view-request"
                                                    ]
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "6|3Kgs2B34mqxB22hmHc9W8N2GYXxwg3OTtWT4I4A84756196e"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "operationId": "0d29c740dc06d77602af8095bf4654cc",
                "responses": {
                    "200": {
                        "description": "Response Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "properties": {
                                                "status": {
                                                    "type": "integer",
                                                    "example": 200
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "logout success"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "LoginRequest": {
                "title": "Login Request",
                "description": "",
                "properties": {
                    "identifier": {
                        "type": "string",
                        "example": "Joko Widodo"
                    },
                    "password": {
                        "type": "string",
                        "example": "Password123$"
                    }
                },
                "type": "object"
            },
            "RegisterRequest": {
                "title": "Register Request",
                "description": "",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Joko Widodo"
                    },
                    "email": {
                        "type": "string",
                        "example": "jokowi@gmail.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "Password123$"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "example": "Password123$"
                    },
                    "nip_nik": {
                        "type": "string",
                        "example": "199001012020121004"
                    },
                    "phone": {
                        "type": "string",
                        "example": "081234567895"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Sanctum Token"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}